==> src/1_news_cl_pre.ipynb <==
# ニューストピック分類 前処理


## Google ドライブの接続  
以下を実行。表示されるurlからGoogleアカウントにサイン。表示される認証用のコードを以下に表示されるテキストボックスにコピペ。  
→ サインインしたGoogleアカウントのGoogleドライブがマウントされる。

from google.colab import drive
drive.mount("/content/drive/")

## 主な設定

data_path = '/content/drive/My Drive/Colab Notebooks/data'   # 作業用フォルダ

## データセットの読み込み
ロンウイット社が公開している[livedoorニュースのデータセット](https://www.rondhuit.com/download.html#ldcc)を読み込む。

import os
dl_file_path = os.path.join(data_path, 'ldcc-20140209.tar.gz')

! wget -c -P '{data_path}' https://www.rondhuit.com/download/ldcc-20140209.tar.gz
! tar -zxf '{dl_file_path}' -C '{data_path}'

import glob
import pandas as pd
text_data  = []   # 文章のリスト
label_data = []   # ラベルのリスト
cat_counts = []   # カテゴリ別サンプル数
text_path = os.path.join(data_path, "text")
dir_names = [base_name for base_name in os.listdir(path=text_path)
    if os.path.isdir(os.path.join(text_path, base_name))]  # ディレクトリ一覧
for i, dir_name in enumerate(dir_names):
    files = glob.glob(os.path.join(text_path, dir_name, "*.txt"))  # ファイル一覧
    for j, file in enumerate(files):
        if os.path.basename(file) == "LICENSE.txt":
            continue
        with open(file, "r") as f:
            text = f.readlines()[3:]  # 記事を読み込み、見出しを削除する
        text = "".join(text)
        text = text.translate(
            str.maketrans({"\n":"", "\t":"", "\r":"", "\u3000":""}))
            # 簡易的なクレンジング
        text_data.append(text)
        label_data.append(i)
    cat_counts.append(j+1)
text_label_data = pd.DataFrame({'text':text_data, 'label':label_data})
cat_info = pd.DataFrame({
    'label':range(len(dir_names)), 
    'name':dir_names,
    'cat_count': cat_counts})
cat_info

## データセットの分割と確認

from sklearn.model_selection import train_test_split
news_train, news_test = train_test_split(text_label_data, shuffle=True, 
            stratify=text_label_data['label'])  # 訓練用とテスト用に分割
news_train.head()

news_test.head()

## データの保存


news_train.to_csv(os.path.join(data_path, "news_train.csv"), index=False)
news_test .to_csv(os.path.join(data_path, "news_test.csv" ), index=False)
cat_info  .to_csv(os.path.join(data_path, "cat_info.csv"  ), index=False)

## データのロード確認

news_train = pd.read_csv(
    os.path.join(data_path, "news_train.csv"), index_col=False)
news_train.head()

news_test = pd.read_csv(
    os.path.join(data_path, "news_test.csv" ), index_col=False)
news_test.head()

cat_info = pd.read_csv(
    os.path.join(data_path, "cat_info.csv" ), index_col=False)
cat_info
