==> src/2_news_cl_train.ipynb <==
# ニューストピック分類 学習



## Google ドライブの接続  
以下を実行。表示されるurlからGoogleアカウントにサイン。表示される認証用のコードを以下に表示されるテキストボックスにコピペ。  
→ サインインしたGoogleアカウントのGoogleドライブがマウントされる。

from google.colab import drive
drive.mount("/content/drive/")

## 主な設定

data_path = '/content/drive/My Drive/Colab Notebooks/data'   # 作業用フォルダ

## ライブラリのインストール

!pip install transformers==4.10
!pip install datasets==1.11
!pip install fugashi==1.1
!pip install ipadic==1.0

## ModelとTokenizerの読み込み
日本語の事前学習済みモデルと、紐づいたTokenizerを読み込む。  
[cl-tohoku/bert-base-japanese-whole-word-masking](https://huggingface.co/cl-tohoku/bert-base-japanese-whole-word-masking)を使用する。

from transformers import BertForSequenceClassification, BertJapaneseTokenizer
model_name = 'cl-tohoku/bert-base-japanese-whole-word-masking'
sc_model = BertForSequenceClassification.from_pretrained(model_name, num_labels=9)
sc_model.cuda()
tokenizer = BertJapaneseTokenizer.from_pretrained(model_name)

## データセットの読み込み
see. [huggingface datasets.load_dataset method](https://huggingface.co/docs/datasets/package_reference/loading_methods.html#datasets.load_dataset)

import os
from datasets import load_dataset
import pandas as pd
news_train = pd.read_csv(
    os.path.join(data_path, "news_train.csv"), index_col=False)
news_test = pd.read_csv(
    os.path.join(data_path, "news_test.csv" ), index_col=False)

news_train.head()

news_test.head()

## 語数の評価

# 形態素解析
import torch
def tokenize(text):
    words = tokenizer.tokenize(text)
    word_ids = tokenizer.convert_tokens_to_ids(words)  # インデックスに変換
    word_tensor = torch.tensor([word_ids])  # テンソルに変換
    return word_tensor

def count_words(texts):
    words_counts = []
    for text in texts:
        x = tokenize(text)
        words_counts.append(len(x[0]))
    return pd.DataFrame({'#_of_words':words_counts})

count_words(news_train['text']).describe()

count_words(news_test['text']).describe()

## データローダーのセット

train_data = load_dataset("csv", 
    data_files=os.path.join(data_path, "news_train.csv"), split="train")
print(train_data)
test_data  = load_dataset("csv", 
    data_files=os.path.join(data_path, "news_test.csv") , split="train")
print(test_data)

## 形態素解析

MAX_LENGTH = 512
def tokenize(batch):
    return tokenizer(batch["text"], padding=True, 
                     truncation=True, max_length=MAX_LENGTH)

train_data = train_data.map(tokenize, batched=True, batch_size=len(train_data))
train_data.set_format("torch", columns=["input_ids", "label"])

test_data = test_data.map(tokenize, batched=True, batch_size=len(test_data))
test_data.set_format("torch", columns=["input_ids", "label"])

## 評価関数の定義
`sklearn.metrics`を使用。  


from sklearn.metrics import accuracy_score

def compute_metrics(result):
    labels = result.label_ids
    preds = result.predictions.argmax(-1)
    acc = accuracy_score(labels, preds)
    return {
        "accuracy": acc,
    }

## Trainerの設定
see. [Trainer](https://huggingface.co/transformers/main_classes/trainer.html#id1)  
see. [TrainingArguments](https://huggingface.co/transformers/main_classes/trainer.html#trainingarguments)  

from transformers import Trainer, TrainingArguments

training_args = TrainingArguments(
    output_dir = "./results",
    num_train_epochs = 2,
    per_device_train_batch_size = 8,
    per_device_eval_batch_size = 32,
    warmup_steps = 500,  # 学習係数が0からこのステップ数で上昇
    weight_decay = 0.01,  # 重みの減衰率
    evaluation_strategy = "epoch",
    logging_dir = "./logs",
)

trainer = Trainer(
    model = sc_model,
    args = training_args,
    compute_metrics = compute_metrics,
    train_dataset = train_data,
    eval_dataset = test_data,
)

## モデルの訓練（fine tuning）


trainer.train()

## モデルの評価

trainer.evaluate()

## モデルの保存

sc_model.save_pretrained(data_path)
tokenizer.save_pretrained(data_path)

## モデルの読み込み確認


loaded_model = BertForSequenceClassification.from_pretrained(data_path) 
loaded_tokenizer = BertJapaneseTokenizer.from_pretrained(data_path)
