==> src/6_news_cl_test_quantization_ML256_with_CPU.ipynb <==
# ニューストピック分類 量子化 Model（MAX_LENGTH 256）の評価 with CPU



## Google ドライブの接続  
以下を実行。表示されるurlからGoogleアカウントにサイン。表示される認証用のコードを以下に表示されるテキストボックスにコピペ。  
→ サインインしたGoogleアカウントのGoogleドライブがマウントされる。

from google.colab import drive
drive.mount("/content/drive/")

## 主な設定

data_path = '/content/drive/My Drive/Colab Notebooks/data'   # 作業用フォルダ

## ライブラリのインストール

!pip install transformers==4.10
!pip install datasets==1.11
!pip install fugashi==1.1
!pip install ipadic==1.0

import os
import torch
import pandas as pd
from transformers import BertForSequenceClassification, BertJapaneseTokenizer

## モデルの読み込み


loaded_model = BertForSequenceClassification.from_pretrained(data_path) 
# loaded_model.cuda()
loaded_tokenizer = BertJapaneseTokenizer.from_pretrained(data_path)

## Quantization（量子化）

loaded_model = torch.quantization.quantize_dynamic(
    loaded_model, {torch.nn.Linear}, dtype=torch.qint8)

## テストデータの読み込み

cat_info = pd.read_csv(os.path.join(data_path, "cat_info.csv"), index_col=False)
cat_names = cat_info['name']
cat_names

news_test = pd.read_csv(
    os.path.join(data_path, "news_test.csv" ), index_col=False)
news_test.head()

## サンプルデータの予測

# 形態素解析
MAX_LENGTH = 256
def tokenize(text):
    words = loaded_tokenizer.tokenize(text)
    word_ids = loaded_tokenizer.convert_tokens_to_ids(words)  # インデックスに変換
    word_tensor = torch.tensor([word_ids[:MAX_LENGTH]])  # テンソルに変換
    return word_tensor

index = 30
sample_text  = news_test.loc[index, 'text']
sample_label = news_test.loc[index, 'label']
print('text :', sample_text)
print('true label:', sample_label, cat_names[sample_label])

x = tokenize(sample_text)
# x = x.cuda()  # GPU対応
y = loaded_model(x)  # 予測
pred = y[0].argmax(-1).item()  # 最大値のインデックス
print("pred label:", pred, cat_names[pred])

# 精度評価と性能評価

from sklearn.metrics import accuracy_score
from tqdm import tqdm
test_texts  = news_test['text']
test_labels = news_test['label']

%%time
preds = []
for text in tqdm(test_texts):
    x = tokenize(text)
    # x = x.cuda()  # GPU対応
    y = loaded_model(x)  # 予測
    pred = y[0].argmax(-1).item()  # 最大値のインデックス
    preds.append(pred)

print(accuracy_score(test_labels, preds))
